Move IntegrationEnroll/* into Integration/Enroll/*
* Rename AWS into AwsOidc, story expects pascal case
but ran into non-story issues where casing was a problem
even renaming AWS to Aws wasn't enough, i don't know
why

* Capitalize browser directory to Browser
