+++ web/packages/teleport/src/Integrations/Enroll/AwsOidc/IAM/AssignRoleModal.tsx
