+++ web/packages/teleport/src/Integrations/Enroll/AwsOidc/IAM/SharedComponents.tsx
