+++ web/packages/teleport/src/Integrations/Enroll/AwsOidc/browser/Cursor.tsx
