+++ web/packages/teleport/src/Integrations/Enroll/AwsOidc/browser/favicon.png
