+++ web/packages/teleport/src/Integrations/Enroll/AwsOidc/instructions/FirstStageInstructions.tsx
