+++ web/packages/teleport/src/Integrations/Enroll/IntegrationTiles.tsx
