+++ web/packages/teleport/src/Integrations/Enroll/common.tsx
